Fix OLTSWITCH-348/OLTTS-5414 Configuration error on the TG16 when rebooting the SF with the maximum Group List configured
After a rebooting event on the MX a flush of configurations on the Line Cards is triggered. If the group list is already at is max capacity the request for creating new groups will be rejected. Even thought they are the same. To avoid this behaviour we have moved the validation to the place where the group is being created.

svn path=/trunk/; revision=1713