Fix OLTSWITCH-361/OLTTS-5634 MC traffic stops flowing after a long duration test
This issue not only affects the MGMD Lib, but also all threads (e.g. CCMs, LACPs,...) that attempt to send packets. It is only verified when it is attempted to send a packet to a port that has it's link down. What observed so far, is that the SDK has a buffer of ~1 MB for packets transmission. When the buffer is exceed the SDK blocks the access of the callback _tx_dispatch(). This takes place until the buffer is freed. Which does not take place, when a link state is down.
To prevent this behavior two conditions were added inside of the routine dtlPduTransmit(), and by this order:
1)- Return Success If the matrix is slave
2) -Return Success If the port has it's link state down.

svn path=/trunk/; revision=1860